feat: api 수정, 추가

See merge request s08-bigdata-dist-sub2/S08P22E106!12

+++ hast/src/main/java/com/cst/hast/dto/response/CountryArticleResponse.java
@@ -17,7 +17,7 @@ public class CountryArticleResponse {
     private String url;
     private String imgUrl;
     private Integer category;
-    private Long score;
+    private double score;
     private String timeStamp;
     private float latitude;
     private float longitude;
